+++ CHANGELOG.md
@@ -32,12 +32,41 @@ Attention: The newest changes should be on top -->
 
 ### Added
 
+- ENH: Implement optional plot saving [#597](https://github.com/RocketPy-Team/RocketPy/pull/597)
+
+### Changed
+
+-
+
+### Fixed
+
+-
+
+## [v1.6.2] - 2024-11-08
+
+### Added
+
+
 - ENH: Expansion of Encoders Implementation for Full Flights. [#679](https://github.com/RocketPy-Team/RocketPy/pull/679)
+- ENH: Implement optional plot saving [#597](https://github.com/RocketPy-Team/RocketPy/pull/597)
+
+### Changed
+
+-
+
+### Fixed
+
+-
+
+## [v1.6.2] - 2024-11-08
+
+### Added
+
 - ENH: add structural to total mass ratio for motor and rocket [#713](https://github.com/RocketPy-Team/RocketPy/pull/713)
 
 ### Changed
 
-
+- REL: bumps up rocketpy version to v1.6.2 [#724](https://github.com/RocketPy-Team/RocketPy/pull/724)
 
 ### Fixed